Separated dialog and dialog container
